requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Flask-Babel==3.0.0
pytz==2022.7
Flask-SocketIO==5.3.6
reportlab==4.0.4
pandas==2.3.2 --only-binary :all:
openpyxl==3.1.5
gunicorn==21.2.0
gevent==25.4.1
Cython==0.29.36  # تثبيت نسخة محددة من Cython لتجنب مشاكل compile
qrcode[pil]>=7.4.2
requests>=2.28.0
Flask-Caching>=2.1.0
redis>=5.0.0